Revert separate useSaveDraft hook
